requirements-test.txt
fastapi
pydantic
SQLAlchemy
pytest
httpx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
asgiproxy
prometheus_fastapi_instrumentator
python-dateutil
feedgen
feedparser
itsdangerous